chore: Update npm dependencies and fix connection error
This commit updates the npm dependencies in the `package.json` file. It adds the `mongoose` package with version `8.3.4` and updates the `next` package to version `14.2.3`. It also adds the `@types/mongoose` package with version `5.11.97` to the devDependencies. Additionally, it updates the `eslint` package to version `8` and the `eslint-config-next` package to version `14.2.3`.

In addition, this commit fixes a connection error in the `connectDB.ts` file where the process was exiting prematurely.